+++ tests/testthat/test-test-Hs.R
@@ -56,20 +56,3 @@ testthat::test_that("test org.Hs.eg.db", {
       c("", "", "", "org.Hs.egSYMBOL2EG"))
    
 })
-
-testthat::test_that("test org.Mm.eg.db", {
-   # skip if org.Hs.eg.db is not available
-   testthat::skip_if_not_installed("org.Mm.eg.db")
-   
-   ann_lib_m <- "org.Mm.eg.db";
-   testthat::expect_equal(
-      freshenGenes(tvalues, ann_lib=ann_lib_m)$SYMBOL,
-      c("", "", "", "Ccl8", "Ccl8"))
-
-   testthat::expect_equal(
-      freshenGenes(tvalues,
-         ignore.case=TRUE,
-         ann_lib=ann_lib_m)$SYMBOL,
-      c("Ccl8", "Adck1", "Adck1,Ccl8", "Ccl8,Mcpt2", "Adck1"))
-   
-})